[dashboard] Update versions for GrimoireLab tools
This code updates the versions of GrimoireLab and Kidash used
in CROSSMINER. GrimoireLab version is set to 0.1.4, while Kidash
one is set to 0.4.12 (instead of using the master branch version).
This change is needed to fix incompatibility issues that may appear
between the Kidash dependency used by GrimoireLab (0.1.4) and the
Kidash master branch version (currently 0.4.13).